requirements.txt
boto3==1.18.60
openai==0.11.0
python-dotenv==0.19.5
streamlit==1.5.0